+++ src/components/Marks/Label.vue
@@ -1,7 +1,6 @@
 <script>
 import Mark from '../../mixins/Marks/Mark.js'
-import { textAnchorPoint } from '../../utils/anchorPoint.js'
-import createSVGStyle from '../../mixins/Marks/utils/createSVGStyle.js'
+import { renderSVG } from '../../rendering/label.js'
 
 export default {
   mixins: [Mark],
@@ -91,34 +90,6 @@ export default {
       return renderSVG(createElement, this.$$transform, this._props)
     }
   }
-}
-
-export function renderSVG (createElement, $$transform, props) {
-  let [cx, cy] = $$transform([props.x, props.y])
-  let anchorPoint = textAnchorPoint(props.anchorPoint)
-  let transform = calcTransform(props.rotation, cx, cy)
-  let styles = createSVGStyle(props)
-
-  styles['fontSize'] = props.fontSize + 'px'
-  styles['font-family'] = props.fontFamily
-  styles['font-weight'] = props.fontWeight
-
-  return createElement('text', {
-    attrs: {
-      'x': cx,
-      'y': cy,
-      'fill': props.color,
-      'text-anchor': anchorPoint.textAnchor,
-      'dominant-baseline': anchorPoint.dominantBaseline,
-      'transform': transform,
-      'class': 'vgg-label'
-    },
-    style: styles
-  }, props.text)
-}
-
-function calcTransform (rotation, cx, cy) {
-  return `rotate(${rotation}, ${cx}, ${cy})`
 }
 
 </script>
