--- src/mixins/Marks/utils/index.js
@@ -1,13 +1,0 @@
-import parseAesthetic from './parseAesthetic.js'
-import parseCoordinate from './parseCoordinate.js'
-import parseCoordinateSet from './parseCoordinateSet.js'
-import parseGeometry from './parseGeometry.js'
-import parseProperty from './parseProperty.js'
-
-export {
-  parseAesthetic,
-  parseCoordinate,
-  parseCoordinateSet,
-  parseGeometry,
-  parseProperty
-}
